test: merge test files into a single file
